Sumcheck integration (#977)
* refactor code

* refactor

* add fix error handling

* fix unreachable path error handling

* add README

* fix clippy

* fix comments

* fix typo

* fix clippy

* Update README.md

---------
