+++ provers/sumcheck/src/prover.rs
@@ -1,9 +1,12 @@
+use crate::Channel;
+use lambdaworks_crypto::fiat_shamir::default_transcript::DefaultTranscript;
 use lambdaworks_math::field::element::FieldElement;
+use lambdaworks_math::field::traits::HasDefaultTranscript;
 use lambdaworks_math::field::traits::IsField;
 use lambdaworks_math::polynomial::{
     dense_multilinear_poly::DenseMultilinearPolynomial, Polynomial,
 };
-
+use lambdaworks_math::traits::ByteConversion;
 /// Prover for the Sum-Check protocol using DenseMultilinearPolynomial.
 pub struct Prover<F: IsField>
 where
@@ -42,4 +45,51 @@ where
         self.current_round += 1;
         univar
     }
+}
+
+pub fn prove<F>(
+    poly: DenseMultilinearPolynomial<F>,
+) -> (FieldElement<F>, Vec<Polynomial<FieldElement<F>>>)
+where
+    F: IsField + HasDefaultTranscript,
+    <F as IsField>::BaseType: Send + Sync,
+    FieldElement<F>: ByteConversion,
+{
+    let mut prover = Prover::new(poly);
+    let claimed_sum = prover.c_1();
+    let mut transcript = DefaultTranscript::<F>::default();
+    let n = prover.poly.num_vars();
+    let mut proof_polys = Vec::with_capacity(n);
+
+    let univar = prover.poly.to_univariate();
+    proof_polys.push(univar.clone());
+
+    transcript.append_felt(&claimed_sum);
+    for coeff in &univar.coefficients {
+        transcript.append_felt(coeff);
+    }
+
+    // Get first challenge
+    let mut challenge = transcript.draw_felt();
+
+    // Subsequent rounds
+    for round in 0..n - 1 {
+        // Adjust challenge to include round number to avoid replay attacks
+        let r_j = &challenge + FieldElement::<F>::from(round as u64);
+
+        // Execute round and get next univariate polynomial
+        let univar = prover.round(r_j.clone());
+        proof_polys.push(univar.clone());
+
+        // Only generate next challenge if this isn't the final round
+        if round < n - 2 {
+            let intermediate_sum = univar.evaluate(&r_j);
+            transcript.append_felt(&intermediate_sum);
+            for coeff in &univar.coefficients {
+                transcript.append_felt(coeff);
+            }
+            challenge = transcript.draw_felt();
+        }
+    }
+    (claimed_sum, proof_polys)
 }
